use bigint for budget

+++ api/src/backend/db/migrations/202210061410_create_brief.ts
@@ -14,7 +14,7 @@ export async function up(knex: Knex): Promise<void> {
 
         // in months atm.  
         builder.integer("duration");  
-        builder.integer("budget");          
+        builder.bigInteger("budget");          
         builder.integer("experience_id");
 
         // stored in its own table
